pyproject.toml
[project]
name = "promptworks-backend"
version = "0.1.0"
description = "Prompt management and evaluation backend."
authors = [{ name = "PromptWorks Team" }]
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "fastapi[standard]>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.7.0",
    "pydantic-settings>=2.2.1",
    "sqlalchemy>=2.0.0",
    "alembic>=1.13.0",
    "psycopg[binary]>=3.1.0",
    "redis>=5.0.0",
    "celery>=5.3.0",
    "httpx>=0.27.0",
    "python-multipart>=0.0.9",
    "pillow>=10.0.0",
]

[project.optional-dependencies]
dev = [
    "ruff>=0.5.0",
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.10.0",
    "poethepoet>=0.27.0"
]

[build-system]
requires = ["setuptools>=67.0"]
build-backend = "setuptools.build_meta"

# 显式限定 setuptools 仅打包 app 目录，避免 flat-layout 带来的包发现冲突。
[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

# 默认启用 dev 依赖组，确保本地开发命令（如 pytest-cov、poethepoet）可用。
[tool.uv]
default-groups = ["dev"]

[dependency-groups]
dev = [
    "ruff>=0.5.0",
    "pytest>=8.2.0",
    "pytest-asyncio>=0.23.0",
    "pytest-cov>=4.1.0",
    "mypy>=1.10.0",
    "poethepoet>=0.27.0"
]

# 配置 pytest 默认启用覆盖率统计并要求最低 90% 覆盖率。
[tool.pytest.ini_options]
addopts = "-s -v --cov=app --cov-report=term-missing --cov-fail-under=90"

# 使用 PoeThePoet 统一管理开发流程任务，方便快速执行格式化、类型检查和测试。
[tool.poe]
# 如需查看所有任务，可运行 `poe -h`。

[tool.poe.executor]
type = "simple"

[tool.poe.env]
UV_CACHE_DIR = ".uv_cache"
UV_NO_SYSTEM_RESOLUTION = "1"
UV_OFFLINE = "0"
UV_DYNAMIC_STORE_DISABLED = "1"

[tool.poe.tasks.format]
help = "使用 Ruff 对后端关键目录执行代码格式化，保持风格统一。"
cmd = "uv run python -m ruff format app alembic tests"

[tool.poe.tasks.lint]
help = "使用 MyPy 对 app 与 tests 目录进行静态类型检查，提前发现潜在类型错误。"
cmd = "uv run mypy app tests"

[tool.poe.tasks.test]
help = "运行 pytest 并输出覆盖率统计（最低 90% 覆盖率）。"
cmd = "uv run pytest"

[tool.poe.tasks."test-all"]
help = "顺序执行格式化、类型检查与测试，确保提交前的基本质量验证。"
sequence = ["format", "lint", "test"]

[tool.poe.tasks.server]
help = "通过 uv 启动 FastAPI 开发服务器，便于本地调试接口。"
cmd = "uv run fastapi dev app/main.py"

[tool.poe.tasks."frontend"]
help = "启动前端"
shell = "cd ./frontend && npm run dev"

# Ruff 配置，避免格式化命令扫描缓存目录导致权限异常。
[tool.ruff]
src = ["app", "tests"]
extend-exclude = [
    ".venv",
    ".uv_cache",
    ".cache",
    "frontend/node_modules",
    "frontend/dist"
]
